requirements.txt
# Core dependencies
web3==6.11.3
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional dependencies
requests==2.31.0
pandas==2.1.4
numpy==1.25.2 